Fix for amd64: use int instead of long int for 32 bit values.

+++ lua-gtk/src/hash-jenkins.c
@@ -8,8 +8,8 @@
  */
 
 
-typedef  unsigned long  int  ub4;   /* unsigned 4-byte quantities */
-typedef  unsigned       char ub1;   /* unsigned 1-byte quantities */
+typedef  unsigned int  ub4;   /* unsigned 4-byte quantities */
+typedef  unsigned char ub1;   /* unsigned 1-byte quantities */
 
 #define hashsize(n) ((ub4)1<<(n))
 #define hashmask(n) (hashsize(n)-1)
